pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inote2obsidian"
version = "0.1.0"
description = "Apple Notes to Obsidian incremental sync service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
inote2obsidian = "inote2obsidian.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
